+++ baselines/split_net/mlp_split_policy.py
@@ -6,7 +6,7 @@ from baselines.common.distributions import make_pdtype
 import numpy as np
 
 # the same as mlp policy for now
-# placeholder if a special mlp policy is needed
+# placeholder if a special mlp policy isn
 class MlpSplitPolicy(object):
     recurrent = False
     def __init__(self, name, *args, **kwargs):
@@ -21,7 +21,7 @@ class MlpSplitPolicy(object):
         sequence_length = None
         print(self.pdtype)
 
-        ob = U.get_placeholder(name="ob", dtype=tf.float32, shape=[sequence_length] + list(ob_space.shape))
+        ob = U.get_placeholder(name="ob", dtype=tf.float64, shape=[sequence_length] + list(ob_space.shape))
 
         with tf.variable_scope("obfilter"):
             self.ob_rms = RunningMeanStd(shape=ob_space.shape)
@@ -37,7 +37,7 @@ class MlpSplitPolicy(object):
             last_out = tf.nn.tanh(U.dense(last_out, hid_size, "polfc%i"%(i+1), weight_init=U.normc_initializer(1.0)))
         if gaussian_fixed_var and isinstance(ac_space, gym.spaces.Box):
             mean = U.dense(last_out, pdtype.param_shape()[0]//2, "polfinal", U.normc_initializer(0.01))
-            logstd = tf.get_variable(name="logstd", shape=[1, pdtype.param_shape()[0]//2], initializer=tf.zeros_initializer())
+            logstd = tf.get_variable(name="logstd", shape=[1, pdtype.param_shape()[0]//2], initializer=tf.zeros_initializer(), dtype=tf.float64)
             pdparam = U.concatenate([mean, mean * 0.0 + logstd], axis=1)
         else:
             pdparam = U.dense(last_out, pdtype.param_shape()[0], "polfinal", U.normc_initializer(0.01))
